Ch 12 - Chapter complete. Added individual counters per sex. Also took the chance to correctly swap around incorrectly labeled male/female indicators. First, had to swap around the bar graphs. Second, had to swap the icons on the ending indicators.
I had the wish to keep female as the first sex and male as the second sex. Female = 0, male = 1. That included making sure that the female bar graph appeared to the left of male. Female above the male counter. And all the correct icon usage where male is a circle and female is a triangle.